pyproject.toml
[build-system]
requires = ["setuptools>=75.1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "JobFunnel"
dynamic = ["version"]
description = "Automated tool for scraping job postings."
readme = "readme.md"
requires-python = ">=3.11"
license = {text = "MIT License"}
authors = [
    { name = "Paul McInnis", email = "paulmcinnis99@gmail.com" },
    { name = "Bradley Kohler" },
    { name = "Jose Alarcon" },
    { name = "Erich Mengore" },
    { name = "Mark van der Broek" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "beautifulsoup4>=4.6.3",
    "lxml>=4.2.4",
    "requests>=2.19.1",
    "python-dateutil>=2.8.0",
    "PyYAML>=5.1",
    "scikit-learn>=0.21.2",
    "nltk>=3.4.1",
    "scipy>=1.4.1",
    "selenium>=3.141.0",
    "webdriver-manager>=2.4.0",
    "Cerberus>=1.3.2",
    "tqdm>=4.47.0",
    "python-jobspy",
]

[project.optional-dependencies]
dev = [
    "pytest>=5.3.1",
    "pytest-mock>=3.1.1",
    "pytest-cov",
    "flake8",
    "isort>=5.10.1",
    "black>=24.8.0",
    "pre-commit>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/PaulMcInnis/JobFunnel"
Documentation = "https://github.com/PaulMcInnis/JobFunnel/docs"
Repository = "https://github.com/PaulMcInnis/JobFunnel"
Issues = "https://github.com/PaulMcInnis/JobFunnel/issues"

[project.scripts]
funnel = "jobfunnel.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "jobfunnel.__version__"}

[tool.setuptools.packages.find]
exclude = ["tests", "docs", "images"]

[tool.setuptools]
include-package-data = true

[tool.pytest.ini_options]
addopts = "--cov=jobfunnel"

[tool.black]
line-length = 88
include = '\.pyi?$'  # This will include all .py and .pyi files

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["beautifulsoup4", "lxml", "requests", "python-dateutil", "PyYAML", "scikit-learn", "nltk", "scipy", "selenium", "webdriver-manager", "Cerberus", "tqdm"]
known_first_party = ["jobfunnel"]
default_section = "THIRDPARTY"
force_sort_within_sections = true
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 120
extend-ignore = "E203"